backend_eval/requirements.txt
fastapi
uvicorn
psycopg2-binary
pydantic
python-multipart
PyPDF2
asyncpg
python-docx
openai-whisper


